Update resource files and improve _GameInfo.cshtml
- Removed "Comeback" entry and added "Spawning" and "Spectators" entries in resource files.
- Imported `System.Globalization` in `_GameInfo.cshtml`.
- Introduced `ProgressBarWidth` method in `_GameInfo.cshtml`.
- Replaced "Comeback" with "ScoreDifference" in `_GameInfo.cshtml`.
- Updated progress bar width calculations to use `ProgressBarWidth`.
- Localized strings for "Spawning", "Dead", "Damage", "Health", and "Spectators" in `_GameInfo.cshtml`.
- Improved spectators display logic and format in `_GameInfo.cshtml`.
- Added server IP connect button section in `_GameInfo.cshtml`.
- Changed `CurrentProgress` type in `Scoreboard.cs` from `float` to `decimal`.
- Changed `Progress` type in `Survivor.cs` from `float` to `decimal`.

+++ L4D2PlayStats.Web/L4D2PlayStats/GameInfo/Models/Scoreboard.cs
@@ -6,7 +6,7 @@ public class Scoreboard
     public int InfectedScore { get; set; }
     public int Bonus { get; set; }
     public int MaxBonus { get; set; }
-    public float CurrentProgress { get; set; }
+    public decimal CurrentProgress { get; set; }
     public int CurrentProgressPoints { get; set; }
     public decimal BonusPercentage => MaxBonus == 0 ? 0 : Bonus / (decimal)MaxBonus;
     public int Difference => SurvivorScore - InfectedScore;
